backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
pymongo==4.6.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
bcrypt==4.1.2
tzdata==2024.2
requests==2.31.0
